+++ src/com/scs/stetech1/entities/ServerPlayersAvatar.java
@@ -16,7 +16,7 @@ public abstract class ServerPlayersAvatar extends AbstractPlayersAvatar implemen
 
 	private AbstractGameServer server;
 
-	public ServerPlayersAvatar(IEntityController _module, int _playerID, IInputDevice _input, int eid, byte side) {
+	public ServerPlayersAvatar(IEntityController _module, int _playerID, IInputDevice _input, int eid, int side) {
 		super(_module, _playerID, _input, eid, side);
 
 		server = (AbstractGameServer)_module;
